refactor(authentication): use module-tools
