+++ modules/authentication/src/Authentication.ts
@@ -1,11 +1,8 @@
 import ConduitGrpcSdk, {
-  ConduitActiveSchema,
-  ConfigController,
   DatabaseProvider,
   GrpcCallback,
   GrpcRequest,
   HealthCheckStatus,
-  ManagedModule,
 } from '@conduitplatform/grpc-sdk';
 import path from 'path';
 import { isNil } from 'lodash';
@@ -15,7 +12,7 @@ import { AdminHandlers } from './admin';
 import { AuthenticationRoutes } from './routes';
 import * as models from './models';
 import { AuthUtils } from './utils';
-import { TokenType } from './constants/TokenType';
+import { TokenType } from './constants';
 import { v4 as uuid } from 'uuid';
 import {
   UserChangePass,
@@ -30,6 +27,11 @@ import { runMigrations } from './migrations';
 import metricsSchema from './metrics';
 import { TokenProvider } from './handlers/tokenProvider';
 import { configMigration } from './migrations/configMigration';
+import {
+  ConduitActiveSchema,
+  ConfigController,
+  ManagedModule,
+} from '@conduitplatform/module-tools';
 
 export default class Authentication extends ManagedModule<Config> {
   configSchema = AppConfigSchema;
