+++ modules/authentication/src/admin/index.ts
@@ -1,14 +1,16 @@
 import ConduitGrpcSdk, {
-  ConduitBoolean,
-  ConduitNumber,
   ConduitRouteActions,
   ConduitRouteReturnDefinition,
-  ConduitString,
-  GrpcServer,
   RouteOptionType,
-  RoutingManager,
   TYPE,
 } from '@conduitplatform/grpc-sdk';
+import {
+  ConduitBoolean,
+  ConduitNumber,
+  ConduitString,
+  GrpcServer,
+  RoutingManager,
+} from '@conduitplatform/module-tools';
 import { UserAdmin } from './user';
 import { ServiceAdmin } from './service';
 import { TeamsAdmin } from './team';
