+++ modules/authentication/src/handlers/oauth2/OAuth2.ts
@@ -1,11 +1,8 @@
 import ConduitGrpcSdk, {
   ConduitRouteActions,
   ConduitRouteReturnDefinition,
-  ConduitString,
-  ConfigController,
   GrpcError,
   ParsedRouterRequest,
-  RoutingManager,
   UnparsedRouterResponse,
 } from '@conduitplatform/grpc-sdk';
 import { isNil } from 'lodash';
@@ -26,6 +23,11 @@ import { TeamsHandler } from '../team';
 import { validateStateToken } from './utils';
 import { IAuthenticationStrategy } from '../../interfaces';
 import { TokenType } from '../../constants';
+import {
+  ConduitString,
+  ConfigController,
+  RoutingManager,
+} from '@conduitplatform/module-tools';
 
 export abstract class OAuth2<T, S extends OAuth2Settings>
   implements IAuthenticationStrategy
