+++ modules/authentication/src/models/Team.schema.ts
@@ -1,4 +1,5 @@
-import { ConduitActiveSchema, DatabaseProvider, TYPE } from '@conduitplatform/grpc-sdk';
+import { DatabaseProvider, TYPE } from '@conduitplatform/grpc-sdk';
+import { ConduitActiveSchema } from '@conduitplatform/module-tools';
 
 const schema = {
   _id: TYPE.ObjectId,
