+++ modules/authentication/src/models/TwoFactorBackUpCodes.schema.ts
@@ -1,4 +1,5 @@
-import { ConduitActiveSchema, DatabaseProvider, TYPE } from '@conduitplatform/grpc-sdk';
+import { DatabaseProvider, TYPE } from '@conduitplatform/grpc-sdk';
+import { ConduitActiveSchema } from '@conduitplatform/module-tools';
 import { User } from './User.schema';
 
 const schema = {
